Switch to using 'jetpacker'
Copied bin/webpack and bin/webpack-dev-server from the generated demo. This code was very different from what Bundler created. As long as I run 'bundle exec jets webpacker:compile' before 'bundle exec jets server', the manifest.json is up to date and the JS code works properly. The fact that 'Jets.webpacker?' returns false is evidently a red herring, as this is also the case for the built-in demo.